+++ src/components/Footer/Footer.js
@@ -1,14 +1,10 @@
 import React from 'react';
 import './Footer.scss';
 
-class Footer extends React.Component {
-  render() {
-    return (
-      <footer className="app-footer">
-        <h2>Footer</h2>
-      </footer>
-    );
-  }
-}
+const Footer = ({ title }) => (
+  <footer className="app-footer">
+    <div className="app-footer-title">{title}</div>
+  </footer>
+);
 
 export default Footer;